### 2025-12-09 #9599 @editorialbot
[PRE REVIEW]: solana-pqzk-fullchain: Full on-chain verification of ZK-STARK proofs and post-quantum signatures on Solana
**Submitting author:** @jotaro-yano (<a href="http://orcid.org/0009-0003-5327-9455">Jotaro Yano</a>)
**Repository:** https://github.com/pqzk-labs/solana-pqzk-fullchain
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.2
