+++ .github/workflows/draft-pdf.yml
@@ -7,7 +7,7 @@ jobs:
     name: Paper Draft
     steps:
       - name: Checkout
-        uses: actions/checkout@v4
+        uses: actions/checkout@v6
 
       - name: Build draft PDF
         uses: openjournals/openjournals-draft-action@master
