### 2025-12-22 8fcc561
Bump bincode from 1.3.3 to 3.0.0
Bumps [bincode](https://github.com/bincode-org/bincode) from 1.3.3 to 3.0.0.
- [Commits](https://github.com/bincode-org/bincode/commits)

---
updated-dependencies:
- dependency-name: bincode
  dependency-version: 3.0.0
  dependency-type: direct:production
  update-type: version-update:semver-major
...

Signed-off-by: dependabot[bot] <support@github.com>
