+++ README.md
@@ -44,54 +44,75 @@ Default: run without Anchor build/deploy using the pinned IDL/types under fixed/
 Prereqs (tested with): Rust 1.88.0, Cargo 1.88.0, Solana CLI 2.2.20 (Agave), Node.js v24.4.0, npm 11.4.2, wasm‑pack 0.13.1
 > **Note:** Tested on Linux, macOS, and Windows Subsystem for Linux (WSL2).  
 > Native Windows without WSL is not supported by solana-cli and may not work.
+ 
+### Steps
 
 0. **Install dependencies**
 
-Before running the demo, install the Node.js packages for the CLI example:  
-```
-npm --prefix examples/cli-chat install (or ~/solana-pqzk-fullchain/examples/cli-chat$ npm install)
-```
+    Install Solana CLI: <https://solana.com/docs/intro/installation>
+    
+    Before running the demo, install the Node.js packages for the CLI example:  
+    
+    ```
+    npm --prefix examples/cli-chat install 
+    ```
+
 
 1. **Connect to devnet & fund your wallet**
-```
-solana config set --url devnet
-solana-keygen new (if you don't have wallet)
-solana airdrop 2
-```
-Wallet locations used by this repo  
-Anchor uses the wallet in Anchor.toml:  
-[provider] wallet = "~/.config/solana/id.json" (default in this repo)  
-Client SDK (examples/cli-chat/src/utils/sdk.ts) resolves the wallet in this order:  
-ANCHOR_WALLET → SOLANA_KEYPAIR → ~/.config/solana/id.json (default)
+
+    If you don't have wallet:
+    ```
+    solana config set --url devnet
+    solana-keygen new
+    solana airdrop 2
+    ```
+
+    If you already have a wallet: 
+    ```
+    solana config set --url devnet
+    solana airdrop 2
+    ```
+
+    Wallet locations used by this repo  
+    Anchor uses the wallet in Anchor.toml:  
+    [provider] wallet = "~/.config/solana/id.json" (default in this repo)  
+    Client SDK (examples/cli-chat/src/utils/sdk.ts) resolves the wallet in this order:  
+    ANCHOR_WALLET → SOLANA_KEYPAIR → ~/.config/solana/id.json (default)
 
 2. **Run the pinned devnet demo (no build/deploy needed)**
-```
-npm --prefix examples/cli-chat run setup (or ~/solana-pqzk-fullchain/examples/cli-chat$ npm run setup)
-npm --prefix examples/cli-chat run keys (or ~/solana-pqzk-fullchain/examples/cli-chat$ npm run keys)
-npm --prefix examples/cli-chat run upload (or ~/solana-pqzk-fullchain/examples/cli-chat$ npm run upload)
-npm --prefix examples/cli-chat run finalize (or ~/solana-pqzk-fullchain/examples/cli-chat$ npm run finalize)
-npm --prefix examples/cli-chat run receive (or ~/solana-pqzk-fullchain/examples/cli-chat$ npm run receive)
-```
-Notes:
-- The SDK reads fixed/idl/stark_pqc_verifier.json and fixed/types/stark_pqc_verifier.ts by default.
-- The devnet program id is taken from the IDL’s metadata.address.
-- The commands above still build client-side crates during npm run setup.
 
-**To use your own deployment (Optional):**
+    ```
+    npm --prefix examples/cli-chat run setup 
+    npm --prefix examples/cli-chat run keys 
+    npm --prefix examples/cli-chat run upload 
+    npm --prefix examples/cli-chat run finalize 
+    npm --prefix examples/cli-chat run receive 
+    ```
+
+    Notes:
[…]
