### 2026-01-06 1000f6e
clarify "Quick start" wallet and expected output
