### 2026-01-08 e6d368f
Add CONTRIBUTING.md
