### 2026-01-10 8a4225a
Add Software design section
