### 2026-01-10 03a3306
Add Research impact statement section
