+++ paper.md
@@ -47,6 +47,10 @@ The software is organized as an end-to-end reference stack for Solana L1: an on-
 
 Build vs. contribute: the cryptographic primitives come from existing standards and libraries; the contribution is the Solana-specific integration and reproducible workflow (program + client + benchmarks) that shows how to run these components under Solana’s execution and transaction constraints.
 
+# Research impact statement
+
+This project provides credible near-term significance as a reproducible baseline for evaluating post-quantum–oriented verification on Solana L1. At the time of writing, we are not aware of many publicly available Solana L1 reference implementations that verify both a STARK proof and an NIST-standard post-quantum signature (SLH-DSA, FIPS 205) fully on-chain in a single end-to-end pipeline. The repository includes a runnable CLI demo and benchmark utilities that record transaction-level compute units for each verification phase, enabling others to reproduce results and compare engineering trade-offs (hashing strategy, memory settings, and I/O chunking). The documented adaptations for the Solana SBF environment are intended to help developers who are considering STARK verification or PQ signatures, and to support future CPI-friendly reuse.
+
 # Acknowledgements
 
 This project builds on the Winterfell STARK ecosystem [@winterfell] and references the NIST post-quantum standards for ML-KEM and SLH-DSA [@FIPS203; @FIPS205]. It also relies on Solana’s public documentation for compute budgeting and transaction constraints [@SolanaFees; @SolanaComputeBudget; @SolanaTxSizeLimit]. No specific financial support was received for this work. The author declares no conflicts of interest.
