### 2026-01-10 cc7f3c7
Add AI usage disclosure section
