+++ MAINTENANCE.md
@@ -1,0 +1,14 @@
+# Maintenance
+
+This repository is a research PoC (not audited). Maintenance is best-effort.
+
+## Maintainer
+- Jotaro Yano (@jotaro-yano)
+
+## Support
+- Bugs / questions / requests: GitHub Issues
+- Security: do not open a public issue. See SECURITY.md
+
+## PRs
+- Please link an Issue.
+- CI must pass.
