+++ ROADMAP.md
@@ -1,0 +1,7 @@
+# Roadmap
+
+This plan may change depending on available time/funding.
+
+- Modular CPI interface: STARK-only, SLH-DSA-only, optional combined flow; update IDL/types.
+- Performance/robustness: reduce CU variance; improve stack/heap and hashing/parsing.
+- Adoption: short CPI integration guide; one small devnet reference example.
