### 2026-01-12 #9599 @danielskatz
@jotaro-yano - thanks for the paper updates. I build the paper to ensure that it was would work, but otherwise, we will continue pausing for about another week or so until we finish determining how to handle our scope change for this submission
