### 2026-01-17 #9599 @nkrusch
👋  Hello @dm-git-acc, @okba14, @ravikanthreddy89, @HardikThakkar94, @octetsplicer

would you be interested in reviewing this submission for JOSS?
