### 2026-01-17 #9599 @ravikanthreddy89
@nkrusch Thank you for tagging me. My primary background is in PKI; Solana expertise is limited. That said, can review the work from PQC and security perspective, if that would be helpful. Also happy to review it more broadly, though it may take some time.
