### 2026-01-17 #9599 @nkrusch
Hi @ravikanthreddy89, perhaps a little more information would be useful.

During a JOSS review, each reviewer works with a checklist. The checklist covers various aspects of the submission, e.g., software functionality, documentation, the paper. I encourage you to have a look at the [checklist], to see what is expected.

The expectation is that each reviewer completes an inspection of the submission. If some checklist item is not initially satisfactory, the author will implement fixes until the submission meets the criteria.

Given this added insight of the process, please let me know if you would be interested in reviewing.

[checklist]: https://joss.readthedocs.io/en/latest/review_checklist.html
