### 2026-01-17 #9599 @ravikanthreddy89
## Review checklist for [@ravikanthreddy89 ](https://github.com/ravikanthreddy89)

- [x] The JOSS [conflict of interest (COI) policy](https://github.com/openjournals/joss/blob/main/COI.md) has been reviewed, and there are no conflicts of interest with reviewing this work, or any perceived conflicts have been waived by JOSS for the purpose of this review.

- [x] The JOSS [code of conduct](https://github.com/openjournals/joss/blob/main/CODE_OF_CONDUCT.md) has been read and will be adhered to.

- [x] **Repository:** The source code for this software is publicly available at https://github.com/pqzk-labs/solana-pqzk-fullchain
- [x] **License:** The repository contains a plain-text LICENSE or COPYING file with the contents of an [OSI-approved](https://opensource.org/licenses/alphabetical) software license.
- [ ] **Contribution and authorship:** The submitting author ([@jotaro-yano](https://github.com/jotaro-yano), Jotaro Yano) has made major contributions to the software, and the list of paper authors appears appropriate and complete.
- [ ] **Substantial scholarly effort:** The submission meets the scope eligibility described in the [JOSS guidelines](https://joss.readthedocs.io/en/pre-2026-submission/submitting.html#substantial-scholarly-effort).
- [ ] **Data sharing:** If the paper contains original data, the data are accessible to reviewers. If no original data are included, this item is satisfied.
- [ ] **Reproducibility:** If the paper contains original results, the results are reproducible by reviewers. If no original results are included, this item is satisfied.
- [ ] **Human and animal research:** If the paper includes original research involving human participants or animals, it complies with the relevant [JOSS policies](https://joss.readthedocs.io/en/pre-2026-submission/policies.html). If no such data are included, this item is satisfied.

- [ ] **Installation:** Installation proceeds as outlined in the documentation.
- [ ] **Functionality:** The functional claims of the software have been confirmed.
- [ ] **Performance:** Any stated performance claims have been confirmed. If no performance claims are made, this item is satisfied.

- [x] **Statement of need:** The authors clearly state the problems the software is designed to solve and identify the target audience.
- [ ] **Installation instructions:** Dependencies are clearly listed and are ideally managed via an automated package management solution.
- [ ] **Example usage:** Examples demonstrating how to use the software are provided, ideally addressing real-world use cases.
- [ ] **Functionality documentation:** The core functionality of the software is documented to a satisfactory level (e.g., API documentation).
- [ ] **Automated tests:** Automated tests are provided, or manual verification steps are clearly described.
- [ ] **Community guidelines:** Clear guidelines exist for third parties to:
  1. Contribute to the software
  2. Report issues or bugs
  3. Seek user or developer support

- [ ] **Summary:** A clear, high-level description of the software’s functionality and purpose is provided for a broad, non-specialist audience.
- [x] **Statement of need:** The paper includes a dedicated “Statement of need” section describing the problem addressed, target audience, and relation to existing work.
- [ ] **State of the field:** The paper discusses how this software compares with other commonly used tools or packages.
- [ ] **Quality of writing:** The paper is well written and does not require significant edits for clarity, structure, or language.
- [ ] **References:** References are complete and appropriately cited, with correct in-text citation syntax.
