### 2026-01-26 #9599 @nkrusch
:wave: @joshiCoding, @Abinashbunty, @amitkumarj441, @gsfyrakis

would you be interested in reviewing this submission for JOSS?
