### 2026-01-26 #9599 @amitkumarj441
> 👋 [@joshiCoding](https://github.com/joshiCoding), [@Abinashbunty](https://github.com/Abinashbunty), [@amitkumarj441](https://github.com/amitkumarj441), [@gsfyrakis](https://github.com/gsfyrakis)
>
> would you be interested in reviewing this submission for JOSS?

@nkrusch I am available from mid-Feb to review this submission. Let me know if that works.
