### 2026-01-26 #9599 @nkrusch
@amitkumarj441 -- yes, that is a reasonable timeline.

Currently, this submission is in pre-review, as I (the editor) am looking for reviewers. Once I confirm enough reviewers, the review phase can actually start. From there, the general recommendation is for reviewers to do an inspection of the submission within ~4 weeks. This means that even in the fastest case, we are looking at about the month of February as the window for initial round of reviews. The JOSS review are iterative (authors implement fixes based on feedback), so the timeline to completion always depends on the submission.

Let me know how this sounds. If you accept, I will add you to the list of reviewers.
