### 2026-02-01 #9923 @nkrusch
:wave:  @ravikanthreddy89, @amitkumarj441 and @Abinashbunty - Thank you for agreeing to review this submission! This is the review thread for the paper. All of our communications will happen here from now on.

### Reviewer Instructions

as a new comment in this thread. The checklist contains the JOSS requirements. As you go over the submission, check any items that you feel have been satisfied.

The [JOSS review &nearr;](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) is different from most other journals. Our goal is to work with the authors to help them meet our criteria instead of merely passing judgment on the submission. The reviewers are encouraged to submit issues and pull requests on the software repository. When doing so, please mention `openjournals/joss-reviews/issues/9923` so that a link is created to this thread (and I can keep an eye on what is happening). Also feel free to comment and ask questions on this thread. In my experience, it is better to post comments/questions/suggestions as you come across them instead of waiting until you've reviewed the entire package.

We aim for reviews to be completed within about 4 weeks. Please let me know if any of you require some more time. We can also use EditorialBot (our bot) to set automatic reminders if you know you'll be away for a known period of time.
