### 2026-02-08 pqzk-labs/solana-pqzk-fullchain#64 @ravikanthreddy89
[JOSS Review][Question] Native nodejs crypto API
@jotaro-yano ,

While checking if the Node.js [native crypto API](https://nodejs.org/dist/v24.8.0/docs/api/crypto.html#asymmetric-key-types) supports SLH-DSA, noticed that support was added in v24.8.0. Since this repository is already targeting v24.4, is there a reason to avoid using the native crypto implementation and instead wire the cryptographic primitives via WASM?

OpenSSL versoin 3.5.0+ is required for SLH-DSA support and [it was added in NodeJS v24.5](https://github.com/nodejs/node/blob/main/doc/changelogs/CHANGELOG_V24.md#upgrade-to-openssl-35), but  API support was added in [v24.8.0](https://github.com/nodejs/node/blob/main/doc/changelogs/CHANGELOG_V24.md#other-notable-changes)

(https://github.com/openjournals/joss-reviews/issues/9923)
