+++ .github/workflows/draft-pdf.yml
@@ -16,7 +16,7 @@ jobs:
           paper-path: paper.md
 
       - name: Upload
-        uses: actions/upload-artifact@v6
+        uses: actions/upload-artifact@v7
         with:
           name: paper
           path: paper.pdf
