### 2026-03-28 f35a163
Add State of the field and reorganize paper sections
