### 2026-03-28 24c86e4
Mention ECDSA and EdDSA in State of the field
