+++ paper.md
@@ -25,7 +25,7 @@ bibliography: paper.bib
 
 # Statement of need
 
-Public blockchains preserve artifacts indefinitely, which creates a “record now, break later” risk profile for protocols whose security relies on assumptions that may be weakened by future capabilities, including Shor-type quantum attacks on discrete-log based systems [@Shor1999]. In relation to existing work, practical deployments often favor verifier-efficient proof systems, while this software targets a Solana-specific, post-quantum-oriented end-to-end workflow. A fuller comparison with related approaches and component libraries is given in the following State of the field section.
+Public blockchains preserve artifacts indefinitely, which creates a “record now, break later” risk profile for protocols whose security relies on assumptions that may be weakened by future capabilities, including Shor-type quantum attacks on discrete-log based systems [@Shor1999]. This concern is relevant not only to transaction authentication, but also to proof systems used to verify on-chain computation. In relation to existing work, practical deployments often favor verifier-efficient proof systems, while this software targets a Solana-specific, post-quantum-oriented end-to-end workflow. A fuller comparison with related approaches and component libraries is given in the following State of the field section.
 
 Solana is a useful platform for studying this feasibility because it exposes explicit transaction compute accounting and strict runtime constraints (compute-unit limits, bounded stack, explicit heap-frame requests, and transaction/instruction size limits) that directly shape what can be verified on L1 [@SolanaFees; @SolanaComputeBudget; @SolanaTxSizeLimit]. However, implementing a full verifier pipeline on Solana L1 is non-trivial: hashing costs dominate, stack pressure can trigger SBF failures, memory allocation must be controlled, and large inputs require DoS- and fee-aware streaming.
 
