### 2026-03-28 b61aa2d
Refine State of the field comparison
