### 2026-04-14 pqzk-labs/solana-pqzk-fullchain#86 @ravikanthreddy89
[JOSS Review][Repo][Comments#1]
While the current analysis provides resource usage data for stateless PQ operations, migration to PQ-safe accounts will likely require a stateful transaction that explicitly binds/commits a PQ public key to an account.

Evaluating the resource impact of such stateful transactions would be more informative than stateless uploads, where the public key is always submitted to the network without being associated with account identity.

Additionally, would it be feasible to adapt the client to operate in a stateless mode? If this has already been explored, documenting the associated challenges would provide valuable guidance for future efforts in this direction.

Ref: (https://github.com/openjournals/joss-reviews/issues/9923)
