### 2026-04-14 pqzk-labs/solana-pqzk-fullchain#86 @jotaro-yano
@ravikanthreddy89

Thank you for this helpful comment.

The current implementation focuses on the message-verification flow in this repository and does not currently include a persistent PQ-account binding mechanism. As a result, the current evaluation does not fully reflect the cost of stateful PQ-account migration.

I will make this limitation explicit in the paper. I also opened follow-up issue #87 to explore the challenges and design considerations involved in moving the client toward a more stateless design.
