### 2026-04-19 pqzk-labs/solana-pqzk-fullchain#86 @jotaro-yano
@ravikanthreddy89

Thank you again for this helpful comment.
I revised the paper to make this limitation explicit.
