### 2026-07-12 e4efdf5
Add benchmark images to README
