### 2026-07-12 55d4293
Clarify proof limitation
