### 2026-07-14 df2e205
Update the paper
