### 2026-07-14 ebb5e9f
Update the paper
