+++ paper.md
@@ -25,14 +25,14 @@ bibliography: paper.bib
 
 # Statement of need
 
-Public blockchains preserve artifacts indefinitely, which creates a “record now, break later” risk profile for protocols whose security relies on assumptions that may be weakened by future capabilities, including Shor-type quantum attacks on discrete-log based systems [@Shor1999]. This concern is relevant not only to transaction authentication, but also to proof systems used to verify on-chain computation. In relation to existing work, practical deployments often favor verifier-efficient proof systems, while this software targets a Solana-specific, post-quantum-oriented end-to-end workflow. A fuller comparison with related approaches and component libraries is given in the following State of the field section.
+Public blockchains preserve artifacts indefinitely, which creates a “record now, break later” risk profile for protocols whose security relies on assumptions that may be weakened by future capabilities, including Shor-type quantum attacks on discrete-log based systems [@Shor1999]. This concern is relevant not only to transaction authentication, but also to proof systems used to verify on-chain computation. In relation to existing work, practical deployments often favor verifier-efficient proof systems, while `solana-pqzk-fullchain` targets a Solana-specific, post-quantum-oriented end-to-end workflow. A fuller comparison with related approaches and component libraries is given in the following State of the field section.
 
 Solana is a useful platform for studying the feasibility of fully on-chain verification of STARK proofs and post-quantum signatures because it exposes explicit transaction compute accounting and strict runtime constraints (compute-unit limits, bounded stack, explicit heap-frame requests, and transaction/instruction size limits) that directly shape what can be verified on L1 [@SolanaFees; @SolanaComputeBudget; @SolanaTxSizeLimit]. However, implementing a full verifier pipeline on Solana L1 is non-trivial: hashing costs dominate, stack pressure can trigger SBF failures, memory allocation must be controlled, and large inputs require denial-of-service (DoS) and fee-aware streaming.
 
 `solana-pqzk-fullchain` addresses this gap by providing a complete, reproducible software stack focused on **engineering practicality** rather than proposing a new proof system:
 
 - **Cross-program invocation (CPI)-friendly on-chain verifier surface:** a Solana program that verifies an **SLH-DSA (FIPS 205) signature** and a **Winterfell STARK proof** in separate phases to enable early rejection of malformed or unauthorized payloads [@FIPS205; @winterfell; @pqzkFullchain].
-- **Binding of proofs to application artifacts:** the verifier derives public inputs from `SHA256(ciphertext)` and verifies a minimal AIR (affine counter) as a baseline mechanism for binding a proof to the uploaded ciphertext [@pqzkFullchain].
+- **Binding of proofs to application artifacts:** the verifier derives public inputs from the SHA-256 hash of the ciphertext and verifies a minimal AIR (affine counter) as a baseline mechanism for binding a proof to the uploaded ciphertext [@pqzkFullchain].
 - **DoS- and fee-aware streaming uploads:** payloads are uploaded in bounded chunks with fixed-offset appends and a rolling SHA-256 hash chain, so invalid uploads can be rejected with constant work per chunk [@pqzkFullchain].
 - **Runtime-aware adaptations for Solana SBF:** patched components route SHA-256 hashing through Solana’s `hashv` system call, suppress inlining in Fast Reed–Solomon Interactive Oracle Proofs of Proximity (FRI) hotspots to respect stack limits, and use a bump allocator synchronized to the requested heap frame for predictable memory behavior [@pqzkFullchain].
[…]
