+++ .github/workflows/ci.yml
@@ -38,7 +38,7 @@ jobs:
 
       # Setup Node.js (for npm scripts in examples/cli-chat)
       - name: Setup Node.js
-        uses: actions/setup-node@v6
+        uses: actions/setup-node@v7
         with:
           node-version: '24.4.0'
 
